gcloud-skyscanner/src/CF-Skyscanner-LoopTable/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-pubsub
datetime